# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AuthOut": {
        "description": "Login/register/refresh response: user + token pair.",
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "user",
          "accessToken",
          "refreshToken"
        ],
        "title": "AuthOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginIn": {
        "description": "POST /auth/login body.",
        "properties": {
          "email": {
            "maxLength": 255,
            "minLength": 3,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginIn",
        "type": "object"
      },
...
      "RefreshIn": {
        "description": "POST /auth/refresh body.",
        "properties": {
          "refreshToken": {
            "maxLength": 512,
            "minLength": 16,
            "title": "Refreshtoken",
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ],
        "title": "RefreshIn",
        "type": "object"
      },
      "RegisterIn": {
        "description": "POST /auth/register body.",
        "properties": {
          "email": {
            "maxLength": 255,
            "minLength": 3,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterIn",
        "type": "object"
      },
...
      "RevokeOthersOut": {
        "description": "POST /auth/sessions/revoke-others response.",
        "properties": {
          "revoked": {
            "title": "Revoked",
            "type": "integer"
          }
        },
        "required": [
          "revoked"
        ],
        "title": "RevokeOthersOut",
        "type": "object"
      },
      "SessionItemOut": {
        "description": "One live session of the caller (GET /auth/sessions item).\n\n``expires_in`` is the remaining KV TTL of ``sess:<id>`` in seconds —\nthe same lifetime the refresh token has left.",
        "properties": {
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "expires_in"
        ],
        "title": "SessionItemOut",
        "type": "object"
      },
...
      "SessionsOut": {
        "description": "GET /auth/sessions response: the caller's live sessions.\n\n``current`` is the session id embedded in the caller's own access token\n(``sid`` claim, ADR-0002). ``null`` means the token carried no ``sid``\nclaim, so the current session cannot be pinpointed (revoke-others will\nrefuse with 409 in that case).",
        "properties": {
          "current": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SessionItemOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "current"
        ],
        "title": "SessionsOut",
        "type": "object"
      },
...
      "UserOut": {
        "description": "Public view of a profile.",
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "isSuperadmin": {
            "default": false,
            "title": "Issuperadmin",
            "type": "boolean"
          },
          "isSuspended": {
            "default": false,
            "title": "Issuspended",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Poi Poi Hisab API",
    "version": "0.28.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "description": "Verify credentials and start a session (401 with a uniform message).",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revoke the caller's session and its current refresh token (204).",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "description": "Return the authenticated user's public profile.",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotate a refresh token; reuse of an old token revokes the session.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Create a profile and start a session (201 with a fresh token pair).",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "description": "List the caller's LIVE sessions, newest TTLs included (200).\n\nThe per-user index is pruned of dead members on read, so the response\nonly ever shows sessions that can actually still refresh. ``current`` is\nthe ``sid`` claim of the caller's own access token (ADR-0002); ``null``\nwhen the token carries no ``sid`` claim (current session unknown).",
        "operationId": "list_sessions_api_v1_auth_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions/revoke-others": {
      "post": {
        "description": "Revoke every live session of the caller EXCEPT the current one.\n\nSame teardown semantics as ``POST /auth/logout`` per victim session:\n``sess:<sid>`` + its current ``rt:<hash>`` die together, and the sid is\nremoved from the profile's index. Idempotent — with nothing else live it\nanswers ``{\"revoked\": 0}``. 409 when the caller's token has no ``sid``\nclaim (nothing is revoked: refusing beats guessing what to keep).",
        "operationId": "revoke_other_sessions_api_v1_auth_sessions_revoke_others_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeOthersOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Other Sessions",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}